le debugueur

+++ src/main/java/dao/DaoVille.java
@@ -2,6 +2,6 @@ package dao;
 
 import model.Ville;
 
-public interface DaoVille extends DaoGeneric<Ville,Integer>{
+public interface DaoVille extends DaoGeneric<Ville,Long>{
 
 }
